requirements.txt
discord.py>=2.3.0
PyNaCl>=1.5.0
mcstatus>=11.0.0
yt-dlp>=2024.1.0
Pillow>=10.0.0
openai>=1.0.0
flask>=3.0.0
